Changed IfProperty to take a simple expression
I'm planning to add a new CodeStep class giving developers one way to write code within a sequence. To that end I'm changing things like IfProperty to take a simple string expression.

Currently you can give it the path to a property to check for "truthy" and optionally prefix that property with a "!" to NOT the expression.