Added new ChangeCollection, ChangeList, and IfNotProperty steps
The new ChangeCollection and ChangeList steps are intended to facilitate working with lists and collections in a bot. For the Todo Bot sample they eliminated the need for custom CodeStep's all together.

The new IfNotProperty step is simply the opposite of IfProperty. I identified the need to add it after helping design work through a couple of sample flows.