Added global turn entities collection to DialogContextState class
After consulting with @chrimc we've decided to simplify the way you find entities by making them globaly available along side the conversation state.  The difference is a set of entities only live for a single turn so if you want them to live beyound that you have to save them to some otehr spot in memory.