Extract method on luisClient.ts
Changes:
Refactor (extract method) for simplification, increased reusability and readability.

Now in the methods:
- `predictionResolvePost`
- `predictionResolveGet`

There is an invocation of the new private function `getLocalURL`, passing the appId as the only parameter.

This function is very small, and takes care of using the encodeURIComponent() function internally, returning the local URL ready to use.