Fix issue with state manager not being clonned (#1439)
We were losing the configured memory scopes when AdaptiveDialog tried to convert a DialogContext into a SequenceContext.  These changes fix the issue by moving the task of cloning a DC into the DC class itself.

I also cleaned up the SequenceContext by adding a couple of static methods to seperate the logic for creating vs. cloning a SequenceContext.