[Adaptive] selectors and conditions (#1414)
* [Adaptive] add selectors

* [Adaptive] replace reference to IExpressionParser to ExpressionParserInterface

* [Adaptive] conditions refactored to align with donet implementation

* replaced null with [] on empty arrays

* save turn.dialogEvent to context state

* convert AdaptiveDialogState from interface to class

* use memory snapshot to evaluate expression

* export selectors

* bug fix

* fixed samples

* add tests for conditions

* AdaptiveDialogState is an interface

* remove async suffix on async functions

* removed evil null

* small adjustments

* Fix issue with state manager not being clonned (#1439)

We were losing the configured memory scopes when AdaptiveDialog tried to convert a DialogContext into a SequenceContext.  These changes fix the issue by moving the task of cloning a DC into the DC class itself.

I also cleaned up the SequenceContext by adding a couple of static methods to seperate the logic for creating vs. cloning a SequenceContext.

* rolled back to use user memory scope in samples