feat: port several runtime changes (#3446)
* port: ConfigurationResourceProvider

This ports the ConfigurationResourceProvider from .NET. In addition, it makes some changes to Configuration handling
throughout the runtime.

The .NET configuration interface is not async, so let's match that. In particular, anything that operates on a
Configuration instance can, itself, resolve any async operations before or after mutations.

This will provide better interop with existing Adaptive classes, the prime example being the Resource Explorer.
Unfortunately, most of that code was written in a way that is not terribly async friendly. It's easier to push async
operations outside the Configuration interface, so that's what we do.

* port: coreBotAdapter skill EOC

Fixes https://github.com/microsoft/botbuilder-js/issues/3442

* fix: use turnContext.adapter

* fix: lint error